+++ src/borg/archiver/benchmark_cmd.py
@@ -218,7 +218,7 @@ class BenchmarkMixIn:
 
     def build_parser_benchmarks(self, subparsers, common_parser, mid_common_parser):
 
-        from .common import process_epilog
+        from ._common import process_epilog
 
         benchmark_epilog = process_epilog("These commands do various benchmarks.")
